UMemPool.c: new unit over its certain core (0x8000A194-0x8000B058), 6 sweeps folded; 0x8000AE94 is fabs
The 8-byte function at 0x8000AE94 was split as src/MSL/atanf.c, a file harvested from another
project's MSL source that happens to compile to the same two instructions. It sits inside EA's
UMemPool.c, far from the MSL code (atan is at 0x8015EE8C, in s_atan.c): the unit and the file go
(nothing else used them; CREDITS.md updated), and the function is written in UMemPool.c.
It is fabs, not fabsf: fn_8000AD9C calls it and rounds the result to a float (frsp), and Swing.c's
calls match only with a double return. Renamed in symbols.txt, engine.h and its three callers
(GoGolfCam.c, Golfer.c, Swing.c); no score changes.

+++ src/GoGolfCam.c
@@ -1069,10 +1069,10 @@ u8 fn_800C4520(View* pView, int nPlayer, f32 fUp, f32 fDown) {
     f32 dx = gPlayers[nPlayer].vTarget[0] - gPlayers[nPlayer].ball.vPos[0];
     f32 dy = gPlayers[nPlayer].vTarget[1] - gPlayers[nPlayer].ball.vPos[1];
     f32 fSlope;
-    if (fabsf(dx) < 1e-6f) {
+    if (fabs(dx) < 1e-6f) {
         return 0;
     }
-    fSlope = fabsf(dy / dx);
+    fSlope = fabs(dy / dx);
     if (dy > 0.0f) {
         if (fSlope < fUp) {
             return 0;
